movable options vec to hashset

+++ src/chess/coordinate.rs
@@ -1,13 +1,13 @@
 use super::error::Error;
 
-#[derive(Debug, PartialEq)]
+#[derive(Debug, PartialEq, Clone, Copy, Hash, Eq)]
 pub struct Coordinate {
-    x: u8,
-    y: u8,
+    x: u32,
+    y: u32,
 }
 
 impl Coordinate {
-    pub fn new(x: u8, y: u8) -> Result<Coordinate, Error> {
+    pub fn new(x: u32, y: u32) -> Result<Coordinate, Error> {
         if x < 1 || x > 8 || y < 1 || y > 8 {
             Err(Error::OutOfRange)
         } else {
@@ -15,11 +15,11 @@ impl Coordinate {
         }
     }
 
-    pub fn x(&self) -> u8 {
+    pub fn x(&self) -> u32 {
         self.x
     }
 
-    pub fn y(&self) -> u8 {
+    pub fn y(&self) -> u32 {
         self.y
     }
 }
